refactor(modo-email): split EmailTemplate and MailTransport into separate files
Moves `EmailTemplate`/`TemplateProvider` from `template/mod.rs` into a
new `template/email_template.rs`, and moves `MailTransport` trait and
`transport()` factory from `transport/mod.rs` into separate
`transport/trait_def.rs` and `transport/factory.rs` files, keeping all
`mod.rs` files as imports-only per convention. Bumps to v0.1.2.
